refactor: lazy load source-foreign related modules/aliases
