+++ xonsh/environ.py
@@ -27,7 +27,6 @@ from xonsh.ansi_colors import (
 )
 from xonsh.built_ins import XSH
 from xonsh.codecache import run_script_with_cache
-from xonsh.dirstack import _get_cwd
 from xonsh.events import events
 from xonsh.lib.lazyasd import LazyBool, lazyobject
 from xonsh.platform import (
@@ -42,6 +41,7 @@ from xonsh.platform import (
 from xonsh.tools import (
     DefaultNotGiven,
     DefaultNotGivenType,
+    _get_cwd,
     adjust_shlvl,
     always_false,
     always_true,
