+++ xonsh/xaliases/__init__.py
@@ -1,0 +1,5 @@
+"""
+- Theses executable aliases are loaded on demand using LazyAlias, thus startup time will be improved.
+- Create a separate module for each CLI.
+- Avoid directly importing modules in this package and use as aliases instead.
+"""
